requirements.txt
numpy==1.17.0
Pillow==6.1.0
pyopengl==3.1.3b2
